Uploaded RMD LogReg Takehome
From Januaray midterm, Jeff Rouder 203b Logistic Regression take home exam. Involves importing choice data in a visual illusion inhibition task where the purpose for this assignment is to look at how the odds of pressing A (rather than H) differs based on the orientation of the focus shape (Target), and the letter in the background (Either A or H).